third_party/WebKit/LayoutTests/inspector/sources/debugger/async-callstack-in-console-expected.txt

CONSOLE ERROR: line 23: Uncaught Error: foo
Tests asynchronous call stacks printed in console.

Set timer for test function.
Console was cleared async-callstack-in-console.html:10

    console.trace() async-callstack-in-console.html:16
    timeout1 async-callstack-in-console.html:16
    setTimeout (async)
    testFunction async-callstack-in-console.html:11

    Uncaught Error: foo async-callstack-in-console.html:23
    timeout2 async-callstack-in-console.html:23
    setTimeout (async)
    timeout1 async-callstack-in-console.html:17
    setTimeout (async)
    testFunction async-callstack-in-console.html:11

    console.trace() async-callstack-in-console.html:28
    timeout3 async-callstack-in-console.html:28
    setTimeout (async)
    timeout2 async-callstack-in-console.html:22
    setTimeout (async)
    timeout1 async-callstack-in-console.html:17
    setTimeout (async)
    testFunction async-callstack-in-console.html:11
